chapter5_sample_code/pyproject.toml
[project]
name = "chapter4-sample-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.185.0",
    "mcp[cli]>=1.13.1",
    "openai>=2.6.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
